updated purge so id dont accept 1 and others than numbers

+++ Commands/purge.js
@@ -3,6 +3,8 @@ exports.run = (client, msg, args) => {
   if(!msg.member.hasPermission(8192)) return msg.reply("*You need a role that provide the right to Delete/Manage Messages!*")
   if(!msg.deletable) return msg.reply("I have no rights to delete Messages!")
   let messagecount = parseInt(args.join(' '), 10);
+  if(isNaN(messagecount)) return msg.reply("This Command only accept numbers!")
+  if(messagecount <= 1) return msg.reply("You must purge more than 1 Message!")
   if(messagecount > 100) return msg.reply("You can only delete 100 Messages at the time!")
   msg.channel.fetchMessages({'limit': messagecount}).then(messages => {
       msg.channel.bulkDelete(messages)
